Virtual world added to allow the camera to follow the player. Points system added.

+++ HAPI_Start/Button.cpp
@@ -16,7 +16,7 @@ Button::Button(Vector3& p, unsigned int w, unsigned int h, std::string m, HAPISP
 //Returns true if the mouse is over the bounds of the button
 bool Button::HasClick(Vector3& p)
 {
-	Rectangle buttonRect(m_screenPosition.GetX(), m_screenPosition.GetX() + m_width, m_screenPosition.GetY(), m_screenPosition.GetY() + m_width);
+	Rectangle buttonRect((int)m_screenPosition.GetX(), (int)m_screenPosition.GetX() + m_width, (int)m_screenPosition.GetY(), (int)m_screenPosition.GetY() + m_width);
 
 	if (p.GetX() > buttonRect.m_left && p.GetX() < buttonRect.m_right && p.GetY() > buttonRect.m_top && p.GetY() < buttonRect.m_bottom) {
 		return true;
@@ -34,7 +34,7 @@ void Button::Render(Visualisation& v)
 	}
 
 	v.RenderDefault(m_screenPosition, m_width, m_height, m_colour);
-	HAPI.RenderText(m_screenPosition.GetX(), m_screenPosition.GetY(), HAPISPACE::HAPI_TColour::BLACK, m_message);
+	HAPI.RenderText((int)m_screenPosition.GetX(), (int)m_screenPosition.GetY(), HAPISPACE::HAPI_TColour::BLACK, m_message);
 }
 
 void Button::SetIsActive(bool a)
