+++ HAPI_Start/StatBar.cpp
@@ -1,7 +1,7 @@
 #include "StatBar.h"
 #include "Visualisation.h"
 
-StatBar::StatBar(Vector3& p, float m, HAPISPACE::HAPI_TColour& c)
+StatBar::StatBar(Vector3& p, unsigned int m, HAPISPACE::HAPI_TColour& c)
 {
 	m_screenPosition = p;
 	m_maxValue = m;
@@ -22,8 +22,8 @@ void StatBar::SetValue(unsigned int n)
 //Render a default shape with a length based on the value of the slider
 void StatBar::Render(Visualisation& v)
 {
-	float amountToRender = m_currentValue / m_maxValue;
-	amountToRender *= 10.0f;
+	unsigned int amountToRender = m_currentValue / m_maxValue;
+	amountToRender *= 10;
 
 	v.RenderDefault(m_screenPosition, (unsigned int)amountToRender * 100, 20, m_colour);
 }
